pos_javara: separate saleschannel for pos from account_javara
