+++ pos_javara/__manifest__.py
@@ -1,22 +1,18 @@
 # -*- coding: utf-8 -*-
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 {
-    'name': 'Account Javara',
-    'summary': 'Custom Accounting for Javara',
+    'name': 'PoS Javara',
+    'summary': 'Custom PoS for Javara',
     'license': 'AGPL-3',
     'version': '11.0',
-    'category': 'Accounting',
+    'category': 'Point of Sale',
     'author': 'Arkana, Joenan <joenan@arkana.co.id>',
     'website': 'https://www.arkana.co.id',
-    'description': """Custom Accounting for Javara""",
+    'description': """Custom PoS for Javara""",
     'depends': [
-        'account_reports',
-        'sale',
         'pos_sale',
     ],
     'data': [
-        'wizard/view_aged_partner_wizard.xml',
-        'views/account_view.xml'
     ],
     'demo': [],
     'test': [
